Add serializer class to library and remove from test framework
