+++ src/MalikP. IVAO Library/MalikP.IVAO.Library.Test/Models/GPSDataModelTest.cs
@@ -1,5 +1,5 @@
-﻿using MalikP.IVAO.Library.Models.Other;
-using MalikP.IVAO.Library.Test.Framework;
+﻿using MalikP.IVAO.Library.Common.Extensions;
+using MalikP.IVAO.Library.Models.Other;
 
 using NUnit.Framework;
