+++ src/MalikP. IVAO Library/MalikP.IVAO.Library.Test/Models/GeneralDataModelTest.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
+using MalikP.IVAO.Library.Common.Extensions;
 using MalikP.IVAO.Library.Models.General;
 using MalikP.IVAO.Library.Test.Common;
-using MalikP.IVAO.Library.Test.Framework;
 
 using NUnit.Framework;
