+++ src/MalikP. IVAO Library/MalikP.IVAO.Library.Test/Models/ServerDataModelTest.cs
@@ -1,5 +1,5 @@
-﻿using MalikP.IVAO.Library.Models.Servers;
-using MalikP.IVAO.Library.Test.Framework;
+﻿using MalikP.IVAO.Library.Common.Extensions;
+using MalikP.IVAO.Library.Models.Servers;
 
 using NUnit.Framework;
